Refactor program view
