+++ src/components/Header.vue
@@ -19,8 +19,8 @@
           </b-navbar-nav>
 
           <b-navbar-nav v-if="isLogged">
-            <b-nav-item v-bind:active="menuOption === '/availability'" href="#/availability">Disponibilidad</b-nav-item>
             <b-nav-item v-bind:active="menuOption === '/budgets'" href="#/budgets">Presupuestos</b-nav-item>
+            <b-nav-item v-bind:active="menuOption === '/program'" href="#/program">Programa</b-nav-item>
             <b-nav-item v-bind:active="menuOption === '/grid'" href="#/grid">Grilla</b-nav-item>
           </b-navbar-nav>
